Extract with_server() to a separate module
